+++ modules/bt865.c
@@ -962,11 +962,11 @@ int bt865_command(struct i2c_client *client, unsigned int cmd, void *arg)
 
 	switch(cmd) {
 	case ENCODER_CMD_SETMODE:
-		bt865_setmode((int) arg, client);
+		bt865_setmode((long int) arg, client);
 		bt865_update(client);
 		break;
 	case ENCODER_CMD_ENABLEOUTPUT:
-		data->enableoutput = (int) arg;
+		data->enableoutput = (long int) arg;
 		bt865_update(client);
 		break;
 	default:
